@@ -4,7 +4,7 @@ using Abp.TestBase;
 
 namespace Abp.AspNetCore.TestBase
 {
-    [DependsOn(typeof(AbpTestBaseModule),typeof(AbpAspNetCoreModule))]
+    [DependsOn(typeof(AbpTestBaseModule), typeof(AbpAspNetCoreModule))]
     public class AbpAspNetCoreTestBaseModule : AbpModule
     {
         public override void Initialize()
