@@ -19,9 +19,10 @@ using Microsoft.AspNetCore.Mvc.Razor;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Mvc.RazorPages.Infrastructure;
 using Microsoft.AspNetCore.Mvc.ViewComponents;
-using Microsoft.AspNetCore.Mvc.ViewFeatures.Internal;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json.Serialization;
+using Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation;
+using Microsoft.AspNetCore.Mvc.ViewFeatures;
 
 namespace Abp.AspNetCore
 {
@@ -58,16 +59,12 @@ namespace Abp.AspNetCore
             //Use DI to create view components
             services.Replace(ServiceDescriptor.Singleton<IViewComponentActivator, ServiceBasedViewComponentActivator>());
 
-            //Change anti forgery filters (to work proper with non-browser clients)
-            services.Replace(ServiceDescriptor.Transient<AutoValidateAntiforgeryTokenAuthorizationFilter, AbpAutoValidateAntiforgeryTokenAuthorizationFilter>());
-            services.Replace(ServiceDescriptor.Transient<ValidateAntiforgeryTokenAuthorizationFilter, AbpValidateAntiforgeryTokenAuthorizationFilter>());
-
             //Add feature providers
             var partManager = services.GetSingletonServiceOrNull<ApplicationPartManager>();
             partManager?.FeatureProviders.Add(new AbpAppServiceControllerFeatureProvider(iocResolver));
 
             //Configure JSON serializer
-            services.Configure<MvcJsonOptions>(jsonOptions =>
+            services.Configure<MvcNewtonsoftJsonOptions>(jsonOptions =>
             {
                 jsonOptions.SerializerSettings.ContractResolver = new AbpMvcContractResolver(iocResolver)
                 {
@@ -83,8 +80,8 @@ namespace Abp.AspNetCore
 
             //Configure Razor
             services.Insert(0,
-                ServiceDescriptor.Singleton<IConfigureOptions<RazorViewEngineOptions>>(
-                    new ConfigureOptions<RazorViewEngineOptions>(
+                ServiceDescriptor.Singleton<IConfigureOptions<MvcRazorRuntimeCompilationOptions>>(
+                    new ConfigureOptions<MvcRazorRuntimeCompilationOptions>(
                         (options) =>
                         {
                             options.FileProviders.Add(new EmbeddedResourceViewFileProvider(iocResolver));
