@@ -28,7 +28,8 @@ namespace Abp.AspNetCore.Configuration
 
         public bool UseMvcDateTimeFormatForAppServices { get; set; }
 
-        public List<Action<IRouteBuilder>> RouteConfiguration { get; }
+        public List<Action<IEndpointRouteBuilder>> EndpointConfiguration { get; }
+
 
         public AbpAspNetCoreConfiguration()
         {
@@ -37,7 +38,7 @@ namespace Abp.AspNetCore.Configuration
             DefaultUnitOfWorkAttribute = new UnitOfWorkAttribute();
             ControllerAssemblySettings = new ControllerAssemblySettingList();
             FormBodyBindingIgnoredTypes = new List<Type>();
-            RouteConfiguration = new List<Action<IRouteBuilder>>();
+            EndpointConfiguration = new List<Action<IEndpointRouteBuilder>>();
             IsValidationEnabledForControllers = true;
             SetNoCacheForAjaxResponses = true;
             IsAuditingEnabled = true;
