@@ -42,7 +42,13 @@ namespace Abp.AspNetCore.Configuration
         /// <summary>
         /// Used to add route config for modules.
         /// </summary>
-        List<Action<IRouteBuilder>> RouteConfiguration { get; }
+        //[Obsolete("Use EndpointConfiguration instead !")]
+        //List<Action<IRouteBuilder>> RouteConfiguration { get; }
+
+        /// <summary>
+        /// Used to add route config for modules.
+        /// </summary>
+        List<Action<IEndpointRouteBuilder>> EndpointConfiguration { get; }
 
         AbpControllerAssemblySettingBuilder CreateControllersForAppServices(
             Assembly assembly,
