@@ -135,9 +135,9 @@ namespace Abp.EntityFrameworkCore
                 var filterExpression = CreateFilterExpression<TEntity>();
                 if (filterExpression != null)
                 {
-                    if (entityType.IsQueryType)
+                    if (entityType.IsKeyless)
                     {
-                        modelBuilder.Query<TEntity>().HasQueryFilter(filterExpression);
+                        modelBuilder.Entity<TEntity>().HasQueryFilter(filterExpression);
                     }
                     else
                     {
