@@ -31,9 +31,12 @@ namespace Abp.EntityHistory
         protected virtual Expression<Func<TEntity, bool>> CreateEqualityExpressionForId<TEntity, TPrimaryKey>(TPrimaryKey id)
         {
             var lambdaParam = Expression.Parameter(typeof(TEntity));
+
             var leftExpression = Expression.PropertyOrField(lambdaParam, "Id");
 
-            Expression<Func<object>> closure = () => id;
+            var idValue = Convert.ChangeType(id, typeof(TPrimaryKey));
+
+            Expression<Func<object>> closure = () => idValue;
             var rightExpression = Expression.Convert(closure.Body, leftExpression.Type);
 
             var lambdaBody = Expression.Equal(leftExpression, rightExpression);
