@@ -8,7 +8,7 @@ namespace Abp.EntityFrameworkCore.Tests.Ef
     {
         public DbSet<Blog> Blogs { get; set; }
 
-        public DbQuery<BlogView> BlogView { get; set; }
+        public DbSet<BlogView> BlogView { get; set; }
 
         public DbSet<Post> Posts { get; set; }
 
@@ -33,6 +33,11 @@ namespace Abp.EntityFrameworkCore.Tests.Ef
                         x.Property(p => p.LastAccessTime).HasConversion(new AbpDateTimeValueConverter());
                     });
             });
+
+            modelBuilder
+                .Entity<BlogView>()
+                .HasNoKey()
+                .ToView("BlogView");
 
             base.OnModelCreating(modelBuilder);
         }
