@@ -12,8 +12,8 @@ namespace Abp.Tests.Domain.Entities
             var w1 = new Worker { Id = 5, Name = "Halil ibrahim Kalkan" };
             var w2 = new Worker { Id = 5, Name = "Halil ibrahim Kalkan" };
 
-            Assert.True(w1 == w2, "Same class with same Id must be equal");
-            Assert.True(w1.Equals(w2), "Same class with same Id must be equal");
+            Assert.True(w1.EntityEquals(w2), "Same class with same Id must be equal");
+            Assert.True(w2.EntityEquals(w1), "Same class with same Id must be equal");
 
             Worker w3 = null;
             Worker w4 = null;
@@ -22,16 +22,16 @@ namespace Abp.Tests.Domain.Entities
 
             var m1 = new Manager { Id = 5, Name = "Halil ibrahim Kalkan", Title = "Software Architect" };
 
-            Assert.True(m1 == w1, "Derived classes must be equal if their Ids are equal");
+            Assert.True(m1.EntityEquals(w1), "Derived classes must be equal if their Ids are equal");
 
             var d1 = new Department { Id = 5, Name = "IVR" };
 
-            Assert.False(m1 == d1, "Different classes must not be considered as equal even if their Ids are equal!");
+            Assert.False(m1.EntityEquals(d1), "Different classes must not be considered as equal even if their Ids are equal!");
 
             var w5 = w1;
             w5.Id = 6;
 
-            Assert.True(w5 == w1, "Same object instance must be equal.");
+            Assert.True(w5.EntityEquals(w1), "Same object instance must be equal.");
         }
 
         [Fact]
