@@ -11,39 +11,33 @@ namespace Abp.Tests.Domain.Values
             var address1 = new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42);
             var address2 = new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42);
 
-            Assert.Equal(address1, address2);
-            Assert.Equal(address1.GetHashCode(), address2.GetHashCode());
-            Assert.True(address1 == address2);
-            Assert.False(address1 != address2);
+            Assert.True(address1.ValueEquals(address2));
+            Assert.True(address2.ValueEquals(address1));
         }
 
         [Fact]
         public void Value_Objects_Should_Not_Be_Same_If_Contains_Different_Data()
         {
-            Assert.NotEqual(
-                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42),
-                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5A0"), "Baris Manco Street", 42)
+            Assert.False(
+                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42).ValueEquals(
+                    new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5A0"), "Baris Manco Street", 42))
             );
 
-            Assert.NotEqual(
-                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Streettt", 42),
-                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42)
+            Assert.False(
+                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Streettt", 42).ValueEquals(
+                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42))
             );
 
-            Assert.NotEqual(
-                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42),
-                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 45)
+            Assert.False(
+                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42).ValueEquals(
+                    new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 45))
             );
         }
 
         [Fact]
         public void Value_Objects_Should_Not_Be_Same_If_One_Of_Them_Is_Null()
         {
-            Assert.NotEqual(
-                null,
-                new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42));
-
-            Assert.True(new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42) != null);
+            Assert.False(new Address(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42).ValueEquals(null));
         }
 
         [Fact]
@@ -52,9 +46,8 @@ namespace Abp.Tests.Domain.Values
             var anAddress = new Address2(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42);
             var anotherAddress = new Address2(null, "Another street", 42);
 
-            Assert.NotEqual(anAddress, anotherAddress);
-            Assert.False(anotherAddress.Equals(anAddress));
-            Assert.True(anAddress != anotherAddress);
+            Assert.False(anAddress.ValueEquals(anotherAddress));
+            Assert.False(anotherAddress.ValueEquals(anAddress));
         }
 
         [Fact]
@@ -63,10 +56,8 @@ namespace Abp.Tests.Domain.Values
             var address1 = new Address3(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42);
             var address2 = new Address3(null, "Baris Manco Street", 42);
 
-            Assert.Equal(address1, address2);
-            Assert.Equal(address1.GetHashCode(), address2.GetHashCode());
-            Assert.True(address1 == address2);
-            Assert.False(address1 != address2);
+            Assert.True(address1.ValueEquals(address2));
+            Assert.True(address2.ValueEquals(address1));
         }
 
         [Fact]
@@ -75,8 +66,8 @@ namespace Abp.Tests.Domain.Values
             var address1 = new Address3(new Guid("21C67A65-ED5A-4512-AA29-66308FAAB5AF"), "Baris Manco Street", 42);
             var address2 = new Address3(null, "Another street", 42);
 
-            Assert.NotEqual(address1, address2);
-            Assert.NotEqual(address1.GetHashCode(), address2.GetHashCode());
+            Assert.False(address1.ValueEquals(address2));
+            Assert.False(address2.ValueEquals(address1));
         }
     }
 }
