@@ -203,8 +203,9 @@ namespace Abp.ZeroCore.SampleApp.Core
         public SecurityStampValidator(
             IOptions<SecurityStampValidatorOptions> options,
             SignInManager signInManager,
-            ISystemClock systemClock)
-            : base(options, signInManager, systemClock)
+            ISystemClock systemClock,
+            ILoggerFactory loggerFactory)
+            : base(options, signInManager, systemClock, loggerFactory)
         {
         }
     }
@@ -219,7 +220,8 @@ namespace Abp.ZeroCore.SampleApp.Core
             ILogger<SignInManager<User>> logger,
             IUnitOfWorkManager unitOfWorkManager,
             ISettingManager settingManager,
-            IAuthenticationSchemeProvider schemes
+            IAuthenticationSchemeProvider schemes,
+            IUserConfirmation<User> userConfirmation
         ) : base(
             userManager,
             contextAccessor,
@@ -228,7 +230,8 @@ namespace Abp.ZeroCore.SampleApp.Core
             logger,
             unitOfWorkManager,
             settingManager,
-            schemes)
+            schemes,
+            userConfirmation)
         {
         }
     }
