@@ -1,9 +1,7 @@
-﻿using System;
+using AbpAspNetCoreDemo.Db;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
-using Microsoft.EntityFrameworkCore.Migrations;
-using AbpAspNetCoreDemo.Db;
 
 namespace AbpAspNetCoreDemo.Migrations
 {
