@@ -3,7 +3,7 @@
 @{
     Layout = null;
 }
-
+@Html.AntiForgeryToken()
 <!DOCTYPE html>
 
 <html>
