@@ -1,14 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Abp.AspNetCore.Mvc.RazorPages;
+﻿using Abp.AspNetCore.Mvc.RazorPages;
 using Abp.UI;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace AbpAspNetCoreDemo.Pages
 {
+    [IgnoreAntiforgeryToken]
     public class ExceptionFilterPageDemoModel : AbpPageModel
     {
         public JsonResult OnGet()
