@@ -1,14 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc.RazorPages;
+﻿using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace AbpAspNetCoreDemo.Pages
 {
     public class RazorPageDemo : PageModel
     {
         public void OnGet()
+        {
+
+        }
+
+        public void OnPost()
         {
 
         }
