@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Abp.AspNetCore.Mvc.RazorPages;
-using Abp.UI;
+﻿using Abp.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace AbpAspNetCoreDemo.Pages
 {
+    [IgnoreAntiforgeryToken]
     public class ResultFilterPageDemoModel : AbpPageModel
     {
         public IActionResult OnGet()
